+++ includes/Modules/DirectCheckout/block.asset.php
@@ -1,0 +1,5 @@
+<?php
+return [
+    'dependencies' => ['wp-blocks', 'wp-element', 'wp-block-editor', 'wp-editor'],
+    'version'      => '1.0.0'
+];